requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
huggingface-hub
transformers
torch
faiss-cpu
llama-cpp-python
sentence-transformers
tqdm
